fixup! iop attention is a 16-bit register

+++ src/mame/altos/altos586.cpp
@@ -404,7 +404,7 @@ private:
 	void mmu_violation();
 
 	// IOP interfacing.
-	void iop_attn_w(offs_t offset, u8 data);
+	void iop_attn_w(offs_t offset, u16 data);
 	u8 hostram_r(offs_t offset);
 	void hostram_w(offs_t offset, u8 data);
 	void hiaddr_w(u8 data);
@@ -513,7 +513,7 @@ void altos586_state::iop_io(address_map &map)
 													// 0x60 586T Generate MULTIBUS interrupt
 }
 
-void altos586_state::iop_attn_w(offs_t offset, u8 data)
+void altos586_state::iop_attn_w(offs_t offset, u16 data)
 {
 	m_iop->set_input_line(INPUT_LINE_NMI, ASSERT_LINE);
 	m_iop->set_input_line(INPUT_LINE_NMI, CLEAR_LINE);
